Create Do Something
Do Something repeats macro on all sheets in a workbook. The macro Run Code adds a threshold and max bin qty to a report exported from SSRS. The macro applies conditional formatting to highlight items that have usage over the threshold. The purpose is to identify out-of-spec items and increase their replenishment quantity as needed.